requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.31
asyncpg==0.29.0
firebase-admin==6.5.0
google-cloud-storage==2.17.0
google-cloud-pubsub==2.21.1
google-cloud-bigquery==3.25.0
google-cloud-secret-manager==2.20.0
pydantic==2.8.0
pydantic-settings==2.3.4
httpx==0.27.0
python-multipart==0.0.9
